# business-platform-api/Models/DTOs/ServiceDtos.cs
namespace business_platform_api.Models.DTOs;

public class ServiceListDto
{
    public int Id { get; set; }
    public string Name { get; set; } = string.Empty;
    public string Slug { get; set; } = string.Empty;
    public string? ShortDescription { get; set; }
    public string? PrimaryImageUrl { get; set; }
}

public class ServiceDetailsDto
{
    public int Id { get; set; }
    public string Name { get; set; } = string.Empty;
    public string Slug { get; set; } = string.Empty;
    public string? ShortDescription { get; set; }
    public string? Description { get; set; }
    
    public List<ServiceImageDto> Images { get; set; } = new List<ServiceImageDto>();
}

public class ServiceImageDto
{
    public int Id { get; set; }
    public string ImageUrl { get; set; } = string.Empty;
    public string? AltText { get; set; }
    public bool IsPrimary { get; set; }
    public int DisplayOrder { get; set; }
}